+++ tree/nodes/join_source.go
@@ -1,11 +1,12 @@
 package nodes
 
-// JoinSource is a specific Binary node.
+// JoinSourceNode is a specific BinaryNode.
 type JoinSourceNode struct {
   Left  *RelationNode // Left child of the JoinSource node, a pointer to a Relation.
   Right []interface{} // Right child of the JoinSource node contains joins and their instructions
 }
 
+// JoinSourceNode factory method.
 func JoinSource(relation *RelationNode) *JoinSourceNode {
   source := new(JoinSourceNode)
   source.Left = relation